smallpt - erand() is a Windows problem, not a MSVC one

+++ smallpt/org_single_precision.cpp
@@ -11,7 +11,7 @@
 // position, also color (r,g,b)
 
 #include "classes.h"
-#include "msvc.h"
+#include "win.h"
 #include <vector>
 
 namespace org_sp {
@@ -94,7 +94,7 @@ Vec radiance(const Ray &r, int depth, unsigned short *Xi) {
 }
 inline void compute_inner(int y, int w, int h, int samps, Ray &cam, Vec &cx, Vec &cy, Vec &r, Vec *c) {
 	//fprintf(stderr, "\rRendering (%d spp) %5.2f%%", samps * 4, 100.*y / (h - 1));
-	for(unsigned short x = 0, Xi[3] = { 0, 0, y*y*y }; x < w; x++)   // Loop cols
+	for(unsigned short x = 0, Xi[3] = { 0, 0, (unsigned short)(y*y*y) }; x < w; x++)   // Loop cols
 		for(int sy = 0, i = (h - y - 1)*w + x; sy < 2; sy++)     // 2x2 subpixel rows
 			for(int sx = 0; sx < 2; sx++, r = Vec()) {        // 2x2 subpixel cols
 				for(int s = 0; s < samps; s++) {
